extracting list of houses
